// app/2026/data/types.ts
export interface TeamInStandingsLike {
  teamId: number;
  rank: number | null;
  problemsSolved: number;
  institution: string;
  teamName: string;
}
// ...
export enum TeamRankStatus {
  NONE = "none",
  D3_1 = "from South Pacific and other super-regions",
  D3_2 = "not solving any problem",
  D3_3 = "of winner universities",
  D3_4 = "fourth or lower ranked teams of each university",
}

export interface TeamInFilteredStandingsLike extends TeamInStandingsLike {
  status: TeamRankStatus;
  recalculatedRank: number;
}